add folders Models and Value Objects
